requirements.txt
Flask
Flask-PyMongo
PyJWT
Werkzeug
pymongo
bson
gunicorn
python-dotenv